src/main/java/ru/sfedu/sprintspherepk/util/PropertiesUtil.java
package ru.sfedu.sprintspherepk.util;

import java.io.FileInputStream;
import java.io.IOException;
import java.util.Properties;

public class PropertiesUtil {
    public static Properties loadProperties(String path) {
        Properties properties = new Properties();
        try (FileInputStream fis = new FileInputStream(path)) {
            properties.load(fis);
        } catch (IOException e) {
            System.err.println("Failed to load properties file: " + e.getMessage());
        }
        return properties;
    }
}